a lot of work

+++ SlnProject/SlnKapperszaak/WpfAdmin/Afpbeheren.xaml.cs
@@ -13,43 +13,50 @@ using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
 using MyClassLibrary;
 
-namespace WpfKlant
+namespace WpfAdmin
 {
     /// <summary>
-    /// Logique d'interaction pour afpbeheren.xaml
+    /// Logique d'interaction pour Afpbeheren.xaml
     /// </summary>
-    public partial class afpbeheren : Window
+    public partial class Afpbeheren : Window
     {
-        Gebruiker Klant;
-        public afpbeheren(Gebruiker klantinp)
+        public Afpbeheren()
         {
             InitializeComponent();
-            Klant = klantinp;
-            foreach (MyClassLibrary.Afspraak item in MyClassLibrary.Afspraak.GetAfspraken(Klant))
-            {
-                lbAfspraken.Items.Add(item);
-            }
-
-
+            lbAfspraken.Items.Add("Selecteer een datum");
         }
 
         private void lbAfspraken_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
-            if(lbAfspraken.SelectedIndex != -1)
+            if (lbAfspraken.SelectedIndex != -1)
             {
-                lblDate.Content = lbAfspraken.SelectedIndex;
                 MyClassLibrary.Afspraak afsp = lbAfspraken.SelectedItem as MyClassLibrary.Afspraak;
                 lblDate.Content = $"Tijdstip: {afsp.Datum.DayOfWeek} {afsp.Datum.Day} {afsp.Datum.Month} {afsp.Datum.Year} om {afsp.Tijd.ToString()}";
                 lblKapper.Content = $"Kapper: {afsp.Kpper.Voornaam} {afsp.Kpper.Achternaam}";
+                lblKlant.Content = $"Klant: {afsp.Klant.Achternaam} {afsp.Klant.Voornaam}";
             }
-                
+
         }
 
         private void btnAnnuleren_Click(object sender, RoutedEventArgs e)
         {
-            MyClassLibrary.Afspraak.DeleteAfspraak((MyClassLibrary.Afspraak)lbAfspraken.SelectedItem);
+            if (MessageBox.Show("Bent u zeker dit afspraak te willen annuleren?", "Verwijderen", MessageBoxButton.OKCancel) == MessageBoxResult.OK)
+            {
+                MyClassLibrary.Afspraak.DeleteAfspraak((MyClassLibrary.Afspraak)lbAfspraken.SelectedItem);
+                lbAfspraken.Items.Clear();
+                foreach (Afspraak item in Afspraak.GetAfsprakenByDate((DateTime)CalDatePicker.SelectedDate))
+                {
+                    lbAfspraken.Items.Add(item);
+                }
+            }
+            
+        }
+
+        private void CalDatePicker_SelectedDatesChanged(object sender, SelectionChangedEventArgs e)
+        {
             lbAfspraken.Items.Clear();
-            foreach (MyClassLibrary.Afspraak item in MyClassLibrary.Afspraak.GetAfspraken(Klant))
+
+            foreach (Afspraak item in Afspraak.GetAfsprakenByDate((DateTime)CalDatePicker.SelectedDate))
             {
                 lbAfspraken.Items.Add(item);
             }
